requirements.txt
h5py==3.7.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.5.5.64
Pillow==9.3.0
scipy==1.7.3
setuptools==61.2.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
pytest==7.2.0